requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cryptography==2.7
decorator==4.4.0
ecdsa==0.13.2
Flask==1.1.1
flask-oidc-ex==0.5.5
Flask-WTF==0.14.2
future==0.17.1
httplib2==0.13.1
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.15.1
Jinja2==2.10.1
jwcrypto==0.6.0
MarkupSafe==1.1.1
oauth2client==4.1.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
Pygments==2.4.2
python-jose==3.0.1
python-jwt==3.2.4
python-keycloak==0.17.5
requests==2.22.0
rsa==4.0
six==1.12.0
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
Werkzeug==0.15.6
WTForms==2.2.1
